requirements.txt
# packages
# python 2.7
numpy==1.16.3
tensorflow-gpu==1.2.0